+++ src/test/java/org/dyn4j/dynamics/contact/SolvableContactTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2021 William Bittle  http://www.dyn4j.org/
+ * Copyright (c) 2010-2022 William Bittle  http://www.dyn4j.org/
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are permitted 
@@ -33,7 +33,7 @@ import junit.framework.TestCase;
 /**
  * Tests the methods of the {@link SolvableContact} class.
  * @author William Bittle
- * @version 4.1.0
+ * @version 5.0.1
  * @since 4.1.0
  */
 public class SolvableContactTest {
@@ -66,7 +66,7 @@ public class SolvableContactTest {
 		TestCase.assertEquals(false, c.ignored);
 		TestCase.assertEquals(null, c.r1);
 		TestCase.assertEquals(null, c.r2);
-		TestCase.assertEquals(true, c.isSolved());
+		TestCase.assertEquals(false, c.isSolved());
 		
 		
 	}
@@ -84,9 +84,14 @@ public class SolvableContactTest {
 		SolvableContact c = new SolvableContact(id, p, d, p1, p2);
 		
 		TestCase.assertEquals(false, c.ignored);
-		TestCase.assertEquals(true, c.isSolved());
+		TestCase.assertEquals(false, c.isSolved());
 		
 		c.ignored = true;
+		TestCase.assertEquals(true, c.ignored);
 		TestCase.assertEquals(false, c.isSolved());
+		
+		c.solved = true;
+		TestCase.assertEquals(true, c.ignored);
+		TestCase.assertEquals(true, c.isSolved());
 	}
 }